4. Implemented functionality for services - simple_search
