+++ tests/conftest.py
@@ -1,0 +1,17 @@
+import pytest
+import pandas as pd
+
+
+@pytest.fixture
+def df_for_tests():
+    test_df = pd.DataFrame({
+        "Дата операции": ["30.12.2021 01:00:00", "20.12.2021 01:00:00", "10.12.2021 01:00:00"],
+        "Дата платежа": ["30.12.2021", "20.12.2021", "10.12.2021"],
+        "Сумма операции": [1442.11, -23.22, -334.55],
+        "Сумма операции с округлением": [1442.11, 23.22, 334.55],
+        "Категория": ["ЖКХ", "Бонус", "Маркет"],
+        "Описание": ["ЖКХ Услуги", "Бонус за покупку", "Яндекс Маркет"],
+        "Номер карты": [None, "*2333", "*3222"],
+        "Статус": ["OK", "OK", "OK"]})
+
+    return test_df
